MGMT-16600: rename container package image
