+++ internal/image/testdata/bootc_status.json
